feat: create tool to convert letters to hashtags
Created a script which create elegant words made with hashtags.
The aim of this tool is to create phrases to read titles within
the comments of the scripts in a simple and elegant way.
At the moment in can print: A, B, C, D, E, F, P, e, k, m, u, v and y.